requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.13.0
aiosignal==1.4.0
asgiref==3.10.0
attrs==25.4.0
blinker==1.9.0
click==8.3.0
colorama==0.4.6
Flask==3.1.2
frozenlist==1.8.0
greenlet==3.2.4
h11==0.16.0
idna==3.10
itsdangerous==2.2.0
Jinja2==3.1.6
keyboard==0.13.5
MarkupSafe==3.0.3
multidict==6.7.0
pillow==11.3.0
playwright==1.55.0
propcache==0.4.0
pyee==13.0.0
typing_extensions==4.15.0
Werkzeug==3.1.3
yarl==1.22.0
